telegram_bot/requirements.txt
aiogram==3.4.1
aiohttp==3.9.3
asyncpg==0.29.0
psycopg2-binary==2.9.9
